archive/custom_GPT/netflix_app/requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain_core
langchain_neo4j
langgraph
python-dotenv
ollama